+++ generic-server/src/main/java/org/holodeckb2b/bdxr/smp/server/ui/viewmodels/ParticipantFormData.java
@@ -16,15 +16,11 @@
  */
 package org.holodeckb2b.bdxr.smp.server.ui.viewmodels;
 
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.stream.Collectors;
 
 import org.holodeckb2b.bdxr.common.datamodel.impl.IdentifierImpl;
 import org.holodeckb2b.bdxr.smp.datamodel.impl.IDSchemeImpl;
-import org.holodeckb2b.bdxr.smp.server.datamodel.ServiceMetadataTemplate;
 import org.holodeckb2b.bdxr.smp.server.db.entities.ParticipantEntity;
-import org.holodeckb2b.bdxr.smp.server.db.entities.ServiceMetadataTemplateEntity;
 import org.holodeckb2b.commons.util.Utils;
 
 import lombok.Getter;