+++ generic-server/src/main/java/org/holodeckb2b/bdxr/smp/server/utils/MailSenderUtil.java
@@ -17,7 +17,6 @@
 package org.holodeckb2b.bdxr.smp.server.utils;
 
 import java.util.Collection;
-import java.util.Collections;
 import java.util.Map;
 
 import org.holodeckb2b.commons.util.Utils;