+++ peppol-smp/src/main/java/org/holodeckb2b/bdxr/smp/server/services/peppol/SMLClient.java
@@ -25,6 +25,7 @@ import java.security.NoSuchAlgorithmException;
 import java.security.UnrecoverableKeyException;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
+import java.time.ZonedDateTime;
 import java.util.GregorianCalendar;
 import java.util.function.Function;
 import java.util.logging.Logger;
@@ -68,7 +69,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.ws.client.core.WebServiceTemplate;
 import org.springframework.ws.soap.client.SoapFaultClientException;
 import org.springframework.ws.soap.client.core.SoapActionCallback;
+import org.springframework.ws.transport.http.HttpComponents5ClientFactory;
 import org.springframework.ws.transport.http.HttpComponents5MessageSender;
+import org.springframework.ws.transport.http.HttpComponentsMessageSender;
 
 import ec.services.wsdl.bdmsl.data._1.PrepareChangeCertificateType;
 import jakarta.xml.bind.JAXBElement;
@@ -196,11 +199,14 @@ public class SMLClient implements SMLIntegrationService {
 
 	@Override
 	public void updateSMPCertificate(String smpId, Certificate cert) throws SMLException {
+		ZonedDateTime activationDate = cert.getActivationDate();
+		if (activationDate == null) 
+			throw new SMLException("Certificate activation date must be at least one day in the future");
+		
 		try {
 			PrepareChangeCertificateType certUpdate = new PrepareChangeCertificateType();
 			XMLGregorianCalendar xmlDate = DatatypeFactory.newInstance()
-														  .newXMLGregorianCalendar(GregorianCalendar.from(
-																  					cert.getActivationDate()));
+													  .newXMLGregorianCalendar(GregorianCalendar.from(activationDate));																  					
 			certUpdate.setMigrationDate(xmlDate);
 			certUpdate.setNewCertificatePublicKey(CertificateUtils.getPEMEncoded(cert.getX509Cert()));
 
@@ -370,7 +376,7 @@ public class SMLClient implements SMLIntegrationService {
 				new SMLMessageSender(
 						HttpClients.custom()
 								.setConnectionManager(new BasicHttpClientConnectionManager(socketFactoryRegistry))
-								.addRequestInterceptorFirst(new HttpComponents5MessageSender.RemoveSoapHeadersInterceptor())
+								.addRequestInterceptorFirst(new HttpComponents5ClientFactory.RemoveSoapHeadersInterceptor())
 								.build()
 						));