+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.0.3
+##### 
+### Changed
+* Spring Boot version updated to 3.5.9
+* HB2B generic-utils library updated to 2.0.0
+
+### Fixed
+* Internal Error when removing last SMT binding from Participant [#18](https://github.com/holodeck-b2b/Holodeck-SMP/issues/18)
+* Incorrect validation error message when updating network registration [#20](https://github.com/holodeck-b2b/Holodeck-SMP/issues/20)
+* Internal Error when updating certificate "Now" [#21](https://github.com/holodeck-b2b/Holodeck-SMP/issues/21)
+
 ## 3.0.2
 ##### 2025-11-25
 ### Fixed